finalproject/project/requirements.txt
asgiref==3.8.1
Django==4.2.11
filelock==3.13.4
fsspec==2024.3.1
intel-openmp==2021.4.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
sympy==1.12
tbb==2021.12.0
torch==2.3.0
torchvision==0.18.0
typing_extensions==4.11.0
tzdata==2024.1
validate-email-address==1
